bot/requirements.txt
aiogram>=3.4,<4
aiohttp>=3.9,<4
python-dotenv>=1.0
